+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Align 2.0.0
 
+- Add `constraints` property to `Constraints` type to allow access to all of the constraints created using it
+- Add `activate` parameter to `Constraints` initiliazer to optionally disable automatic activation of constraints
 - Remove `addSubview` family of APIs
 
 ## Align 1.2.1